Audio Updates
Refactors out common state variables from both EnemyFSM classes in a
new abstract class Enemy.  This allows the new EnemyManager class to
maintain a list of enemies and get state information from them.
By doing this, we can know if the  player is being attacked at any time
by any enemy and change the music.

We will also be able to use this information to control the camera in
the future.